Add files via upload
COVID-19 inspired social media website 

Senior web design workshop 1 project 

For this project, I was asked to draw inspiration from the inconveniences I faced during the COVID-19 pandemic. The goal of the project was to create a digital platform that acted as a solution to one or more of these inconveniences. My assigned group members and I suggested a web application that could bring communities together during a time of social distancing restrictions.  I acted as lead developer for Phased, while other members of my group took on the roles of content/SEO management, graphic design, and UI design.